mm: PKSMv0.2: Apply pksm-v0.2-for-linux3.6.0 patch
PKSM uses an algorithm and mechanism that directly handles the Anon-pages
that are created/freed by the linux kernel. It is also transparent for user,
while saving CPU time, because it does not need to traverse VMAs to find
a valid anonymous page.

From: https://code.google.com/p/pksm/downloads/list
Signed-off-by: Luis Cruz <ljc2491@gmail.com>